+++ src/OpenMesh/Tools/Subdivider/Uniform/Sqrt3InterpolatingSubdividerLabsikGreinerT.hh
@@ -109,7 +109,7 @@ namespace Uniform    { // BEGIN_NS_UNIFORM
  * Clement Courbet - clement.courbet@ecp.fr
 */
 
-template <typename MeshType, typename RealType = float>
+template <typename MeshType, typename RealType = double>
 class InterpolatingSqrt3LGT : public SubdividerT< MeshType, RealType >
 {
 public:
